// lib/apis/octant.ts
export interface EpochBudgetItem {
  /** User address */
  address: string;
  /** User budget for given epoch, BigNumber (wei) */
  amount: string;
}
// ...
export interface EpochAllocation {
  /** Donor address */
  donor: string;
  /** Funds allocated by donor for the proposal in WEI */
  amount: string;
  /** Proposal address */
  proposal: string;
}
// ...
export interface EpochStats {
  /** ETH proceeds from staking for the given epoch. */
  stakingProceeds: string;
  /** Effectively locked GLMs for the given epoch */
  totalEffectiveDeposit: string;
  /** Total rewards for the given epoch. */
  totalRewards: string;
  /** Total rewards budget allocated to users rewards */
  individualRewards: string;
  /** The amount needed to cover the Octant's costs */
  operationalCost: string;
  /** Rewards users decided to withdraw for the given epoch. */
  totalWithdrawals?: string;
  /** Matching fund budget coming from patrons. */
  patronsRewards?: string;
  /**
   * Total matched rewards for the given epoch.
   *             Includes matchedRewards from Golem Foundation and patronRewards.
   */
  matchedRewards?: string;
  /** The amount that will be used to increase staking and for other Octant related operations. Includes donations to projects that didn't reach the threshold. */
  leftover?: string;
  /** PPF for the given epoch. It's calculated based on substracting Vanillia Individual Rewards from Individual Rewards Equilibrium. */
  ppf?: string;
  /** Community fund for the given epoch. It's calculated from staking proceeds directly. */
  communityFund?: string;
}